+++ components/sql/src/test/scala/pl/touk/nussknacker/sql/utils/StandaloneProcessTest.scala
@@ -8,8 +8,8 @@ import pl.touk.nussknacker.engine.api.{JobData, ProcessVersion}
 import pl.touk.nussknacker.engine.baseengine.api.runtimecontext.EngineRuntimeContextPreparer
 import pl.touk.nussknacker.engine.graph.EspProcess
 import pl.touk.nussknacker.engine.resultcollector.ProductionServiceInvocationCollector
-import pl.touk.nussknacker.engine.standalone.StandaloneScenarioEngine
-import pl.touk.nussknacker.engine.standalone.StandaloneScenarioEngine.StandaloneResultType
+import pl.touk.nussknacker.engine.requestresponse.RequestResponseEngine
+import pl.touk.nussknacker.engine.requestresponse.RequestResponseEngine.RequestResponseResultType
 import pl.touk.nussknacker.engine.testing.LocalModelData
 import pl.touk.nussknacker.engine.util.SynchronousExecutionContext.ctx
 
@@ -21,7 +21,7 @@ trait StandaloneProcessTest extends Matchers with ScalaFutures {
 
   def contextPreparer: EngineRuntimeContextPreparer
 
-  def runProcess(process: EspProcess, input: Any): StandaloneResultType[List[Any]] = {
+  def runProcess(process: EspProcess, input: Any): RequestResponseResultType[List[Any]] = {
     val interpreter = prepareInterpreter(process)
     interpreter.open()
     try {
@@ -31,8 +31,8 @@ trait StandaloneProcessTest extends Matchers with ScalaFutures {
     }
   }
 
-  private def prepareInterpreter(process: EspProcess): StandaloneScenarioEngine.StandaloneScenarioInterpreter = {
-    val validatedInterpreter = StandaloneScenarioEngine(process,
+  private def prepareInterpreter(process: EspProcess): RequestResponseEngine.RequestResponseScenarioInterpreter = {
+    val validatedInterpreter = RequestResponseEngine(process,
       ProcessVersion.empty, DeploymentData.empty,
       contextPreparer, modelData, Nil, ProductionServiceInvocationCollector, runMode)