+++ engine/base/request-response/api/src/main/scala/pl/touk/nussknacker/engine/requestresponse/api/ResponseEncoder.scala
@@ -1,0 +1,9 @@
+package pl.touk.nussknacker.engine.requestresponse.api
+
+import io.circe.Json
+
+trait ResponseEncoder[-Input] {
+
+  def toJsonResponse(input: Input, result: List[Any]): Json
+
+}