+++ ui/runServer.sh
@@ -10,9 +10,9 @@ SCALA_VERSION=${SCALA_VERSION:-2.12}
 PROJECT_BASE_DIR="../../.."
 #management jars are currently needed to access DeploymentManagers
 FLINK_ENGINE_JAR=$PROJECT_BASE_DIR/engine/flink/management/target/scala-${SCALA_VERSION}/nussknacker-flink-manager.jar
-STANDALONE_ENGINE_JAR=$PROJECT_BASE_DIR/engine/standalone/engine/target/scala-${SCALA_VERSION}/nussknacker-standalone-manager.jar
+REQUEST_RESPONSE_RUNTIME_JAR=$PROJECT_BASE_DIR/engine/base/request-response/runtime/target/scala-${SCALA_VERSION}/nussknacker-request-response-manager.jar
 
-export CLASSPATH="../target/scala-${SCALA_VERSION}/nussknacker-ui-assembly.jar:$FLINK_ENGINE_JAR:$STANDALONE_ENGINE_JAR"
+export CLASSPATH="../target/scala-${SCALA_VERSION}/nussknacker-ui-assembly.jar:$FLINK_ENGINE_JAR:$REQUEST_RESPONSE_RUNTIME_JAR"
 DIST_BASE_DIR="$PROJECT_BASE_DIR/nussknacker-dist/src/universal"
 export CONFIG_FILE="$DIST_BASE_DIR/conf/dev-application.conf"
 export NUSSKNACKER_LOG_LEVEL=DEBUG
@@ -20,7 +20,7 @@ export CONSOLE_THRESHOLD_LEVEL=DEBUG
 
 export MANAGEMENT_MODEL_DIR="$PROJECT_BASE_DIR/engine/flink/management/sample/target/scala-${SCALA_VERSION}"
 export GENERIC_MODEL_DIR="$PROJECT_BASE_DIR/engine/flink/generic/target/scala-${SCALA_VERSION}"
-export STANDALONE_MODEL_DIR="$PROJECT_BASE_DIR/engine/standalone/engine/sample/target/scala-${SCALA_VERSION}"
+export REQUEST_RESPONSE_MODEL_DIR="$PROJECT_BASE_DIR/engine/base/request-response/runtime/sample/target/scala-${SCALA_VERSION}"
 
 export BASE_COMPONENT_DIR="$PROJECT_BASE_DIR/components/base/target/scala-${SCALA_VERSION}"
 export OPENAPI_COMPONENT_DIR="$PROJECT_BASE_DIR/components/openapi/target/scala-${SCALA_VERSION}"